my changes here give you big pp

+++ test6.pl
@@ -7,10 +7,12 @@ use object;
 
 my $x=150;
 #y holds the current y value of the character (up and down)
-my $y=150;
+my $y=50;
 my $velocity=0;
 #below holds the current value for the height of the floor beneath the character
-my $below=150;
+my $below=145;
+#onObject determines if the player is on an object or not
+my $onObject=0;
 #the objects array holds all current visible objects on the screen
 #It is used to cycle through for character collision and all the objects need
 #to be moved left when the character "moves" right
@@ -31,7 +33,18 @@ sub redrawIcon{
 	$id=$canvas->createBitmap($x,$y, -bitmap=>'@Stick.XBM');
 }
 sub gravity{
-	if($y<$below && ($velocity==0 ||$velocity==-1))
+	$onObject=0;
+	foreach my $obj (@objects)								#Sets boolean value base on if player is standing on object
+	{
+		if ($obj->{x1} <=150 && $obj->{y1} >= 150 && ($obj->{y2}-25==$y))
+		{
+			$onObject=1;
+		}
+	}
+	if ($onObject)
+	{
+	}
+	elsif($y<$below && ($velocity==0 ||$velocity==-1))
 	{
 
 		$canvas->delete($id);
@@ -53,7 +66,7 @@ sub move{
 	my ($widget)= @_;
 	my $e=$widget->XEvent;
 	my ($keysym_text, $keysym_decimal) = ($e->K, $e->N);
-	if($keysym_text eq "w" && $y==$below)
+	if($keysym_text eq "w" && ($y==$below || $onObject==1))
 	{
 		$velocity=1;
 		for(1..60)
@@ -66,32 +79,46 @@ sub move{
 		}
 		$velocity=0;
 	}
-	elsif($keysym_text eq "d")
+	elsif($keysym_text eq "d" || $keysym_text eq "a")		#horizontal keys
 	{
+		my $dontMove=0;
+		#first check if the player ran into an obstacle without jumping over
 		foreach my $obj (@objects)
 		{
-			$canvas->delete($obj->{id});
-			$obj->{x1}-=10;	$obj->{y1}-=10;
-			$obj->{id}=$canvas->createRectangle($obj->{x1},$obj->{x2},$obj->{y1},$obj->{y2},-fill=>'green');
-
+			if($keysym_text eq "d")
+			{
+				if($obj->{x1}-5==150 && $obj->{y2}-25<$y)
+					{$dontMove=1;}	
+			}
+			else
+			{
+				if($obj->{y1}+5==150 && $obj->{y2}-25<$y)
+				{$dontMove=1;}
+			}
 		}
-
-	}
-	elsif($keysym_text eq "a")
-	{
-
-		foreach my $obj (@objects)
+		if ($dontMove==0)
 		{
-			$canvas->delete($obj->{id});
-			$obj->{x1}+=10;	$obj->{y1}+=10;
-			$obj->{id}=$canvas->createRectangle($obj->{x1},$obj->{x2},$obj->{y1},$obj->{y2},-fill=>'green');
+			if($keysym_text eq "d")
+			{
+				foreach my $obj (@objects)
+				{
+					$canvas->delete($obj->{id});
+					$obj->{x1}-=5;	$obj->{y1}-=5;
+					$obj->{id}=$canvas->createRectangle($obj->{x1},$obj->{x2},$obj->{y1},$obj->{y2},-fill=>'green');
 
+				}
+			}
+			else
+			{
+				foreach my $obj (@objects)
+				{
+					$canvas->delete($obj->{id});
+					$obj->{x1}+=5;	$obj->{y1}+=5;
+					$obj->{id}=$canvas->createRectangle($obj->{x1},$obj->{x2},$obj->{y1},$obj->{y2},-fill=>'green');
+
+				}
+			}
 		}
-
-	}
-	elsif($keysym_text eq "s")
-	{
-
 	}
 }
